Create AsyncStream.cs
